Login fake do google; Ações de post; Seeder

+++ app/Http/Controllers/PostController.php
@@ -7,14 +7,25 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
+use Illuminate\Support\Facades\DB;
 
 class PostController extends Controller {
 	/**
 	 * Display a listing of the resource.
 	 */
 	public function index(ListPostsRequest $request) {
-		$posts = Post::where('project_id', $request->project_id)->orderBy('created_at', 'desc')->paginate();
+		if($request->user()->projects()->where('id', $request->project_id)->doesntExist()) {
+			return response()->json(['message' => 'Unauthorized'], 403);
+		}
+		//DB::enableQueryLog();
+		$posts = Post::
+			where('project_id', $request->project_id)
+			->with('tags', function ($query) {
+				$query->orderBy('tags.id');
+			})
+			->orderBy('created_at', 'desc')->paginate();
 
+		//dd(DB::getQueryLog());
 		return PostResource::collection($posts);
 	}
 
@@ -33,6 +44,9 @@ class PostController extends Controller {
 	 * Display the specified resource.
 	 */
 	public function show(Post $post) {
+		if(request()->user()->id != $post->user_id) {
+			return response()->json(['message' => 'Unauthorized'], 403);
+		}
 		return new PostResource($post);
 	}
 
@@ -40,7 +54,12 @@ class PostController extends Controller {
 	 * Update the specified resource in storage.
 	 */
 	public function update(UpdatePostRequest $request, Post $post) {
+		if($request->user()->id != $post->project->user_id) {
+			return response()->json(['message' => 'Unauthorized'], 403);
+		}
 		$post->update($request->all());
+		if(isset($request->tags))
+			$post->tags()->sync($request->tags);
 
 		return new PostResource($post);
 	}
@@ -49,8 +68,12 @@ class PostController extends Controller {
 	 * Remove the specified resource from storage.
 	 */
 	public function destroy(Post $post) {
+		if(request()->user()->id != $post->project->user_id) {
+			return response()->json(['message' => 'Unauthorized'], 403);
+		}
 		$post->delete();
 
 		return response()->noContent();
 	}
+
 }
